+++ History.md
@@ -3,6 +3,7 @@ Maven release
 
 1.0.0 / 2013-??-??
 ==================
+  * Remove difference in web-server and file-system instrumentation (https://github.com/tntim96/JSCover/issues/77)
   * Use Semantic Versioning 2.0.0 Guidelines (https://github.com/tntim96/JSCover/issues/76)
   * Reduce branch data JSON size (https://github.com/tntim96/JSCover/issues/73)
   * Add 'Stop Server' Button to UI (https://github.com/tntim96/JSCover/issues/74)